+++ src/main/java/com/homebudget/homeentirety/repositories/TransactionSummaryRepository.java
@@ -3,5 +3,5 @@ package com.homebudget.homeentirety.repositories;
 import com.homebudget.homeentirety.models.TransactionSummary;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransactionSummaryRepository extends JpaRepository<TransactionSummary, Integer> {
+public interface TransactionSummaryRepository extends JpaRepository<TransactionSummary, Long> {
 }
